Manage domains on DRBD failover cluster
The manage-domains script provides a command to evacuate a host and a
command to distribute the running domains to other hosts according to a
configuration file. This depends on a shared storage.

This functionality is currently only available if the DRBD hook script
is installed as well. Patches for other forms of shared storage are
welcome.

The setup of DRBD is not part of this module. This can be done witht the
drbd module from puppetlabs.