fix qemu package name in Debian bullseye
This is similar to #59 in the sense that the newer Debian
release *also* doesn't have a qemu-kvm package, which is now a "meta
package". This matters because as things stand, this module will
constantly be effecting changes on servers, like this:

    Notice: /Stage[main]/Libvirt::Install/Package[qemu-kvm]/ensure: created (corrective)

Instead of creating a one-shot solution, we create data files for the
older release which will, eventually, go away.

One downside of this new package name is that it hardcodes the
architecture. Unfortunately, the package name doesn't match the
architecture provided by facts:

    anarcat@curie:libvirt$ facter -p | grep x86
      hardware => "x86_64",
      platform => "x86_64-linux-gnu",

... so it seems unlikely this can be fixed purely with Hiera logic.