add hiera layer for architecture specific settings
Newest Debian/Ubuntu versions using virtual packages for
qemu-kvm. Since usually you want to install the package
corresponding your architecture we introduce
a architecture specific hiera layer.

To avoid to many files in the root directory, we move
the OS specific packages in directories. (not the family !)

also see #59 and #60