add additional network interface types for domain definition
This is based on the idea of @sebastianrakel and @bastelfreak
in pull request #81. Besides adding bridges this also adds additional
network types.

This deprecates the 'network' and 'portgroup' parameters in favor
of the more flexible 'source' parameter in the network interface hash.
The deprecated parameters will remain functional until next major release.