Revert partly "human readable diff in case of changes; correct sort for portgroups inside XML"
This partly reverts commit 4c559086bde670435a1fd83ba9acfecf2d93f54a (#75).

The part 'human readable diff' is kept, the part 'correct sort...' is reverted.

Fixing the sort in the provider only is the wrong approach. This module uses
the function defined in lib/puppet_x/libvirt/sort_elements.rb to
have equal XML's to compare. The sort_elements function is used in the provider
to ensure sorting of the 'is' part. The 'should' part is sorted using
the function libirt::normalxml on the template XML output.

See Bug #85